Add files via upload
This program comprises the usage of EEPROM memory where the LED can change its state from high to low or low to high with the help of a potentiometer. However, the data of the maximum intensity is saved in the EEPROM memory and the person can pass the value from the Serial Monitor.